line chart update moved to seperate file
